Update several files and create some others
Create a favicon folder in the images folder for better organization and update all the html files to change the place of the favicon elements. Create avis.html file, avis.css file for the properties of the page and avis.js for the stars in the review. Create page_inscription.html file and page_inscription.css file. Add a line to not make the cross appearing when the page is charging in the login.js file.