Add UserRepository
Add app/persistence/UserRepository.py file for the CRUD operations with the DB. Create app/persistence/BaseRepository.py as a base model for CRUD operations with the DB for all the entities. It facilitates the implementation in the entity repository with only specific function as get_by_email for User. Add test file for the UserRepository