Add PrestationRepository
Add app/persistence/PrestationRepository.py. Add tests file for PrestationRepository. Update app/persistence/BaseRepository.py with a function to get entities by attributes. Update all the tests files with this new function. In all entity repository file, add a function that force to use the entity create() function instead of the one of BaseRepository because of the validations inside