Add AppointmentService
Add app/services/AppointmentService.py with the mai_service call to send a notification email to the user and the practitioner. Add the corresponding tests file. Add app/services/mail_service.py which configurates the notification mails to the user and practitioner. Add the corresponding tests file. Add the line 'self.db.session.rollback()' to all the repository files to avoid errors during the creation of the object in the DB. Add integration tests files for all the service files. Add a facade tests file. Launch all the tests. All tests are green for now. Update the readme file adding all the new files in the structure. Update app/config.py adding MAIL_RECIPIENT_PRACTITIONER for the mail configuration.