Debug mode
Update the base_test.py file with debug log to find where the problem was. The problem was utf-8 error and users present in the tables instead of being deleted. Find the solution: each test isolation and clean the db after each test. For the utf-8 problem: the error was from PostgreSQL error message about to many clients written in French and the caracter 'é' was the error. So put max_connexions to 200 instead of 100 and then to 400 because the api tests make a lot of connexion. After the app is over, put the max_connexion back down to 100. Update the test 'test_base_is_clean' in each api tests files. For now, when tests are launch by folder, it all works, except for the api folder because there's too many connexions. Solution: launch tests file one by one. All test are green.