Personal info autofill
Update app/api/v1/users.py adding a new model (user_me_response_model) for the new route /me which get all the user's informations. Update tests files to validate the new route. Update base_files/templates/page_personnelle.html deleting the static value of the fields to be autofilled by calling the /me route. Update base_files/static/js_files/page_personnelle.js to handle the autofill of the fields when the user is loged in. Test on Swagger ok for the recuperation of the user's informations. Test with the front and it works.