Link page_rdv
Update base_files/static/js_files/formualire_rdv.py adding all the logic to load the user data in the form, to load the prestation dynamically in the dropdown menu and to send the appointment form to the DB. Update all the html files with accessibility adding aria-labels. Update css files of the pages linked to the DB with properties for a message alert. Everything is still working, even on Swagger.