Dockerization
Dockerization of the app. Create docker_compose.yml and Dockerfile in the Mots-pour-mauxLM root. Update app.run.py and app/config.py with the new place of the .env. Create a function de_setup() in app/utils.py to launch the docker. Update the readme, still some info to put in.