Add therapie_test database
Add docker_config/init-test-db.sql to create the therapie_test database for tests when launch docker image. Create app/seed.py to create Ghost user and Ghost prestation for reassignment when user or prestation attached to it is deleted. Update entrypoint.sh with the call to seed.py for the creation of the prestations after the tables creation. Update app/tests/base_file.py to launch the tests with all the right paths. Update docker_compose.yml with the new volume for the test database creation. Update the README adding the process with Docker and how to launch it.