Update utils.js
Update app/static/js_files/utils.js adding all the common functions to facilitate the scalability of the code in the js files. Update all the js files adding the utilization of the function isValidInput() for more security. Everything is working but tests file is going to be implementing to tests security fails like XSS, CSRF or SQL injections.