Fix errors on admin page
On admin_user page, when updating a hpone number or an address, it said to fill all the fields whereas it's a PATCH route. Fix the JS code in app/static/js_files/admin_users.js in the function attachUserSaveAndCancelListeners(): separate the optional fields (address and phone number) from the others. Everything is now working well as before.